refactor: RegionCategoryServiceTest에서 연관 관광지 테스트 분리
