+++ backend/turip-app/src/test/java/turip/region/service/RegionCategoryServiceTest.java
@@ -14,13 +14,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import turip.content.repository.ContentRepository;
-import turip.infrastructure.client.KoreaTourismRelatedSpotClient;
-import turip.infrastructure.client.dto.KoreaTourismRelatedSpotResponse.RelatedSpot;
 import turip.region.controller.dto.response.RegionCategoriesResponse;
-import turip.region.controller.dto.response.RelatedSpotsResponse;
 import turip.region.domain.City;
 import turip.region.domain.Country;
-import turip.region.domain.DomesticRegionCategory;
 
 @ExtendWith(MockitoExtension.class)
 class RegionCategoryServiceTest {
@@ -36,9 +32,6 @@ class RegionCategoryServiceTest {
 
     @Mock
     private ContentRepository contentRepository;
-
-    @Mock
-    private KoreaTourismRelatedSpotClient koreaTourismRelatedSpotClient;
 
     @DisplayName("국내 지역 카테고리 조회 시 국내 도시 목록과 기타 카테고리를 반환한다")
     @Test
@@ -152,53 +145,5 @@ class RegionCategoryServiceTest {
         // then
         assertThat(response.regionCategories()).hasSize(1);
         assertThat(response.regionCategories().getFirst().name()).isEqualTo("해외 기타");
-    }
-
-    @DisplayName("지역 카테고리로 연관 관광지를 조회한다")
-    @Test
-    void findRelatedSpotsByRegionCategory1() {
-        // given
-        DomesticRegionCategory category = DomesticRegionCategory.SEOUL;
-
-        RelatedSpot spot1 = mock(RelatedSpot.class);
-        RelatedSpot spot2 = mock(RelatedSpot.class);
-        RelatedSpot spot3 = mock(RelatedSpot.class);
-
-        given(spot1.getRelatedCategoryLargeName()).willReturn("관광지");
-        given(spot1.getRelatedSpotName()).willReturn("북촌한옥마을");
-        given(spot2.getRelatedCategoryLargeName()).willReturn("관광지");
-        given(spot2.getRelatedSpotName()).willReturn("경복궁");
-        given(spot3.getRelatedCategoryLargeName()).willReturn("관광지");
-        given(spot3.getRelatedSpotName()).willReturn("남산타워");
-
-        // 서울의 3개 시군구 코드에 대해 각각 모킹
-        given(koreaTourismRelatedSpotClient.searchRelatedSpots(11, 11110))
-                .willReturn(List.of(spot1));
-        given(koreaTourismRelatedSpotClient.searchRelatedSpots(11, 11440))
-                .willReturn(List.of(spot2));
-        given(koreaTourismRelatedSpotClient.searchRelatedSpots(11, 11140))
-                .willReturn(List.of(spot3));
-
-        // when
-        RelatedSpotsResponse response = regionCategoryService.findRelatedSpotsByRegionCategory(category);
-
-        // then
-        assertAll(
-                () -> assertThat(response.relatedSpots()).hasSize(1),
-                () -> assertThat(response.relatedSpots().getFirst().category()).isEqualTo("관광지"),
-                () -> assertThat(response.relatedSpots().getFirst().spots())
-                        .containsExactlyInAnyOrder("북촌한옥마을", "경복궁", "남산타워")
-        );
-    }
-
-    @DisplayName("'국내 기타' 카테고리는 예외를 던진다")
-    @Test
-    void findRelatedSpotsByRegionCategory2() {
-        // given
-        DomesticRegionCategory category = DomesticRegionCategory.OTHER_DOMESTIC;
-
-        // when & then
-        assertThatThrownBy(() -> regionCategoryService.findRelatedSpotsByRegionCategory(category))
-                .isInstanceOf(turip.common.exception.custom.IllegalArgumentException.class);
     }
 }
